Finalising the tag options
Making the tag options more rebust
also cleaning up some code and renaming the update.php to config.php
also adding the request and response to the admin core controller since we use them often.

Also, the creation of the new post will be done via POST. But the checks will be processed via ajax to give alerts on errors before sending
