+++ App/Controllers/Admin/Config.php
@@ -4,6 +4,7 @@ namespace App\Controllers\Admin;
 
 use App\Models\ConfigModel;
 use Core\AdminController;
+use Core\Traits\StringFunctions;
 
 /**
  * all to do with the config page
@@ -12,6 +13,9 @@ use Core\AdminController;
  */
 class Config extends AdminController
 {
+
+    use StringFunctions;
+
     /**
      * Shows the config page with all of the config options
      * @throws \Twig_Error_Loader
@@ -24,12 +28,40 @@ class Config extends AdminController
         $this->onlyAdmin();
 
 
-
         $configObj = new ConfigModel($this->container);
         $this->data['configList'] = $configObj->getAllConfigOrdered();
-
-
-
         $this->renderView('Admin/Config');
+    }
+
+
+    /**
+     * gets the information in post and sets the corresponding site config details
+     * this is pure php with no ajax
+     * @throws \Exception
+     */
+    public function updateConfig()
+    {
+        //Security checks
+        $this->onlyAdmin();
+        if (!$this->container->getRequest()->isPost()) {
+            $this->alertBox->setAlert('Only post messages allowed', 'error');
+            $this->container->getResponse()->redirect('admin');
+        }
+
+        $configModel = new ConfigModel($this->container);
+        $posts = $this->container->getRequest()->getDataFull();
+        $success = true;
+
+        foreach ($posts as $key => $config) {
+            if (!$configModel->updateConfig($key, $config)) {
+                $success = false;
+            }
+        }
+        if ($success) {
+            $this->alertBox->setAlert('Configuration updates successfully');
+        } else {
+            $this->alertBox->setAlert('error in configuration update', 'error');
+        }
+        $this->container->getResponse()->redirect('admin/config');
     }
 }
