+++ App/Views/Admin/NewPost.twig
@@ -1,88 +1,113 @@
 {% extends "Admin/Base.twig" %}
 
 {% block main %}
-<form action="/admin/new/post" method="post">
+    <form action="/admin/post/create-new-post" method="post" id="newPost">
+        <div class="row">
+            <div class="form-group col-lg-8">
 
-    <div class="row">
-        <div class="form-group col-md-8">
-            <label for="newPostTitle">Post Name</label>
-            <input type="text" class="form-control" id="newPostTitle" name="newPostTitle">
-        </div>
-        <div class="form-group col-md-8">
-            <label for="newPostSlug">Post Slug</label>
-            <input type="text" class="form-control" id="newPostSlug" name="newPostSlug">
-        </div>
-    </div>
-    <div class="row">
-        <div class="form-group col-md-8">
-            <textarea id="newPostTextArea" name="newPostTextArea"></textarea>
-        </div>
-        <div class="form-group col-md-4">
-            <label for="categorySelector">Select a category</label>
-            <select name="categorySelector" id="categorySelector" class="form-control">
-                {% for category in categories %}
-                    <option value="{{ category.idcategories }}">{{ category.category_name }}</option>
-                {% endfor %}
-            </select>
-            <label for="tagSelector">add tags</label>
+                <label for="newPostTitle">Post Name</label>
+                <input type="text" class="form-control" id="newPostTitle" name="newPostTitle">
 
-            <div class="input-group">
+                <label for="newPostSlug">Post Slug</label>
+                <input type="text" class="form-control mb-3" id="newPostSlug" name="newPostSlug">
 
-                <input list="tags" type="text" class="form-control" id="tagSelector" name="tagSelector" data-idtags="" autocomplete="off">
-                <span class="input-group-btn">
+                <textarea id="newPostTextArea" name="newPostTextArea"></textarea>
+
+            </div>
+            <div class="form-group col-lg-4">
+                <label for="categorySelector">Select a category</label>
+                <select name="categorySelector" id="categorySelector" class="form-control">
+                    {% for category in categories %}
+                        <option value="{{ category.idcategories }}">{{ category.category_name }}</option>
+                    {% endfor %}
+                </select>
+
+                <label for="isPublished">Published</label>
+                <select name="isPublished" id="isPublished" class="form-control">
+                    <option value="0">No</option>
+                    <option value="1">Yes</option>
+                </select>
+
+                <label for="isOnFrontPage">On Front Page</label>
+                <select name="isOnFrontPage" id="isOnFrontPage" class="form-control">
+                    <option value="0">No</option>
+                    <option value="1">Yes</option>
+                </select>
+
+                <label for="tagSelector">add tags</label>
+
+                <div class="input-group">
+
+                    <input list="tags" type="text" class="form-control" id="tagSelector" name="tagSelector"
+                           data-idtags="" autocomplete="off">
+                    <span class="input-group-btn">
                     <button class="btn btn-default" type="button" id="addTag">Add</button>
                </span>
-                <datalist id="tags">
-                    {% for tag in tags %}
-                        <option id="tags-{{ tag.tag_name }}" value="{{ tag.tag_name }}" data-idtags="{{ tag.idtags }}"></option>
-                    {% endfor %}
-                </datalist>
+                    <datalist id="tags">
+                        {% for tag in tags %}
+                            <option id="tags-{{ tag.tag_name }}" value="{{ tag.tag_name }}"
+                                    data-idtags="{{ tag.idtags }}"></option>
+                        {% endfor %}
+                    </datalist>
+                </div>
+                <span id="selectedTags"></span>
+
             </div>
-            <span id="selectedTags"></span>
-
         </div>
-    </div>
-
-</form>
+        <button type="submit" class="btn btn-primary ml-3 update-config-button">Create new post</button>
+    </form>
 {% endblock %}
 
 {% block js %}
-<script>
-    setupTinymce("#newPostTextArea");
+    <script>
+        setupTinymce("#newPostTextArea");
 
-    //Slugify the title on key up
-    setupAjax();
-    var $slugRequest = null;
-    $("#newPostTitle").keyup(function () {
-        $text = $("#newPostTitle").val();
-        if($slugRequest !== null){
-            $slugRequest.abort();
+        //Slugify the title on key up
+        setupAjax();
+        var $slugRequest = null;
+        $("#newPostTitle").keyup(function () {
+            $text = $("#newPostTitle").val();
+            if ($slugRequest !== null) {
+                $slugRequest.abort();
+            }
+
+            $slugRequest = $.ajax({
+                url: "/ajax/ajaxSlugify/slugify-string",
+                data: {'slugText-update': $text}
+            }).done(function ($result) {
+                $("#newPostSlug").val($result.slug);
+            });
+        });
+
+        //Add tag
+        var $tagNumber = 0;
+        $("#addTag").on("click", function ($e) {
+            var $value = $("#tagSelector").val().trim();
+            if ($value !== "") { //only do action if we have a tag
+                var $option = $("#tags").find("[value=\"" + $value + "\"]");
+
+                var $tagHtml = "<span>";
+                $tagHtml += "<input name=\"tags[" + $tagNumber + "][name]\" class=\"selected-tags\" type=\"text\" value=\"" + $value + "\" readonly>";
+
+                if ($option.length > 0) { //if the tag already exists, add the id
+                    $tagHtml += "<input name=\"tags[" + $tagNumber + "][id]\" class=\"selected-tags\" type=\"hidden\" value=\"" + $option.data("idtags") + "\" readonly>";
+                }
+
+                $tagHtml += "<i class=\"fa fa-times-circle-o\" aria-hidden=\"true\" onclick=\"closeMe(this)\"></i>";
+                $tagHtml += "</span>";
+                $("#selectedTags").append($tagHtml);
+                $("#tagSelector").val("");
+                $tagNumber += 1;
+            }
+        });
+
+        //remove tag
+        function closeMe($element) {
+            $($element).parent("span").remove();
         }
 
-        $slugRequest = $.ajax({
-            url: "/ajax/ajaxSlugify/slugify-string",
-            data: {'slugText-update': $text}
-        }).done(function($result){
-            $("#newPostSlug").val($result.slug);
-        });
-    });
+        //ajax to do the testing before submit
 
-    //Add tag
-
-    $("#addTag").on("click",function($e){
-        //add the tag + id to the post
-        //Also must create new tags
-        var value = $("#tagSelector").val();
-        var option = $("#tags").find("[value=\""+value+"\"]");
-        var tagHtml = "<span><input class=\"selected-tags\" type=\"text\" data-idtags=\""+option.data("idtags")+"\" value=\""+value+"\" readonly><i class=\"fa fa-times-circle-o\" aria-hidden=\"true\" onclick=\"closeMe(this)\"></i></span>";
-        $("#selectedTags").append(tagHtml);
-        $("#tagSelector").val("");
-    });
-
-    //remove tag
-    function closeMe(element){
-        $(element).parent("span").remove();
-    }
-</script>
+    </script>
 
 {% endblock %}
