+++ Core/AdminController.php
@@ -10,6 +10,19 @@ namespace Core;
 abstract class AdminController extends Controller
 {
     /**
+     * The request object to handle all gets and posts
+     * @var Dependency\Request
+     *
+     */
+    protected $request;
+
+    /**
+     * The response module to handle response messages
+     * @var Dependency\Response
+     */
+    protected $response;
+
+    /**
      * Out placeholders for modules
      * @var object
      */
@@ -22,13 +35,18 @@ abstract class AdminController extends Controller
         $this->loadModules[] = 'AlertBox';
         parent::__construct($container);
 
+        $this->request = $container->getRequest(); //adding our request object as it will be needed in the ajax calls
+        $this->response = $container->getResponse();
+
         //Sending the auth level to the data
         $this->data['userRole'] = $this->auth->getUserRole();
         $this->data['userLevel'] = $this->auth->getUserLevel();
+
     }
 
-    protected function onlyAdmin(){
-        if(!$this->auth->isAdmin()){
+    protected function onlyAdmin()
+    {
+        if (!$this->auth->isAdmin()) {
             $this->alertBox->setAlert("Only admins can access this", 'error');
             $this->container->getResponse()->redirect('/admin');
         }
